requirements.txt
xmltodict
lxml
beautifulsoup4
jsonpath-ng
jinja2
sqlalchemy
psycopg2-binary 
uvicorn
fastapi
fastapi-utils
fastapi_responses~=0.2.1
requests
pytest
sqlalchemy-filters
loguru
pyjwt==2.4.0
fastapi-oidc
python-dotenv
pandas
python-decouple~=3.6
mysqlclient
openai
